moved all functions in his own file.
